requirements.txt
apache-airflow>=2.5.0
great-expectations>=0.15.0
boto3>=1.26.0
pyspark>=3.2.0
delta-spark>=1.1.0
python-dotenv>=0.19.0
requests>=2.26.0
kafka-python>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
dbt-core>=1.0.0
dbt-postgres>=1.0.0

